Simplified search fields and field bridges: - removed many 'keyword', datasource', 'organism' index fields annotations from biopax l3 impl. classes; - added new transient index fields to L3ElementImpl for indexing (paxtools api interfaces are not affected and cannot be used to set this optional fields); - added new private 'uri' setter/getter (for persistence) to BioPAXElementImpl (the 'RDFId' setter impl., which is part of paxtools public api, is not thread-safe); the db table now has 'uri' field instead od 'RDFId'; - cleaned ModelUtils; added/modified several methods; - added useful features to the Fetcher class (this chnages are compatible with prev. revisions); - moved all fetch profile definitions to the one place (L3ElementImpl), renamed.
Technically, the ORM and search index modifications are not backward compatible.
But because the Hibernate/Search is optional (paxtools-core api can be used to create/read/write/analyze biopax without it),
and the core api is not modified, this is still version 4.2.0-SNAPSHOT (not 5.0.0).
This probably affects only cPath2 project at the moment, and is manageable.