Updated SIF creation commands to include optional columns.
Also downgraded the version of the jena library in the json converter to support Java 6. Tests ran with no problems. So the only dependency to Java 7 is at paxtools-search module. While deploying paxtools to maven repositories it is important to remember to build using Java 6, and then separately build paxtools-search using Java 7.